Tracking nodes operations in Nodes Table
This patch introduces nodesInProgress Set in nodesReducer instead of
dataOperationInProgress Boolean. nodesInProgress stores uuids of nodes
which have operation going on them. This advantage is used in nodes
table where a checkbox or loader spinner is used depending on whether
node is in operation or available to trigger an operation.

Change-Id: I89afa8bc3b2c55eed38b923408f4a0d98df1e713