Stop Validation
When Validation is running, user can stop it at any time by clicking
Validation status icon. Doing so updates the workflow execution to
'PAUSED' state. The status of the validation is determined as
new, so user is still obliged to run it later

Validations and Workflow Executions tests added

Implements: blueprint validations-mistral-integration
Change-Id: Ied030d410615e02dbf21777d4be7f2c1124f0db9