Add packaging spec and config files
When a release is made, openstack-infra will produce a tarball of this project.
We add a tarball of the node_modules directory into it via the prepublish hook
in npm (infra uses "npm pack").  This will allow us to build an RPM of this
project without an internet connection.

The RPM spec untars the dependencies, places the directory in the right place
and runs "npm run build" to produce dist files in the dist/ directory.

We're also adding an httpd config file.

Change-Id: Ic685477e11dd28ba5db39752ffe532f518287654