Add packaging spec and config files
When a release is made, openstack-infra will produce a tarball of this
project.  We add a tarball of the node_modules directory into it via the
prepublish hook in npm (infra uses "npm pack").  This will allow us to
build an RPM of this project without an internet connection.

Given the above source tarball, we produce two SRPMs: one for the
project proper, and the other for its dependencies.  Dependencies are
installed into /opt/.

We're also adding an httpd config file.

Change-Id: Ic685477e11dd28ba5db39752ffe532f518287654