Adds basic internationalization support
This patch adds an I18NProvider component to implement react-intl's
translation mechanisms and replaces text strings in the deployment_plan
components with `<FormattedMessage />` instances.

The patch also contains the extracted messages in form of JSON and
`.pot` files (which can be used with zanata.org). See more details in
the README changes.

More components need to be updated once this patch has landed.

Change-Id: Ida9fc65c65bedf377341220be1d7225d1ae58b2c
Implements: blueprint tripleo-ui-i18n-support-for-js