Add dependencies for UI internationalisation
This patch adds new dependencies (and updates one) needed for i18n
features.

`json-loader` and `react-intl` are needed for translation within the app
and will be part of the minimalized JS provided with the package.

`babel-plugin-react-intl` is used as part of the build process to
parse the source and extract messages for translation.

`react-intl-po` is used to convert message files for use with the
zanata service.

Change-Id: I7c5b6e7eb8c72a6cfecb973960277fce9e74723d