Add dependencies for UI internationalisation
This patch adds new dependencies needed for i18n features. All three new
packages are part of the tooling or build process.

The new packages are:

- babel-plugin-react-intl (BSD-3-Clause License)
  https://github.com/yahoo/babel-plugin-react-intl
- json-loader (MIT License)
  https://github.com/webpack/json-loader
- react-intl-po (MIT License)
  https://github.com/evenchange4/react-intl-po

Note: The general change of the `format` metadata fields in the
shrinkwrap file is due to how shrinkwrap handles first vs. subsequent
runs.

Implements: blueprint tripleo-ui-i18n-support-for-js
Change-Id: I7c5b6e7eb8c72a6cfecb973960277fce9e74723d