Relax JSON input validation
Currently the JSON form input validator only passes if the value is a
valid JSON string. This doesn't accout for cases where a paramater
doesn't need to be set. This patch relaxes the validation rule so it
also allows the value to be an empty string.

Change-Id: I4d135120ac3041b223986eef8f54c4de24f20379
Closes-Bug: #1645266
(cherry picked from commit 537618fed2c393cb1fbc0b52f40951dd66e3373c)