Fix blank overcloud credentials
Both the AdminPassword and the PublicVirtualIP are needed to display the
overcloud info. Both are fetched via different Heat API calls, but the
component only checks if one of them is loaded.

This patch adds a selector which combines checking both if
PublicVirtualIP as well as the AdminPassword have been fetched before
removing the loader.

The selector either returns an immutable Map containing the overcloud
credentials or an empty Map, if one of the values is still missing. This
way the same prop can be used for the loader and to display the values.

The stack state has been reworked and stores resource details and the
current stack's environment as separate stack items.

Change-Id: I35ab87ee4a8cb84483ec06f71cb3cee479f8655e
Closes-Bug: #1640538