Fix blank overcloud credentials
Both the AdminPassword and the PublicVirtualIP are needed to display the
overcloud info. Both are fetched via different Heat API calls, but the
component only checks if one of them is loaded.

This patch adds a selector which combines checking both if
PublicVirtualIP as well as the AdminPassword have been fetched before
removing the loader.

Change-Id: I35ab87ee4a8cb84483ec06f71cb3cee479f8655e
Closes-Bug: #1640538