Remove SharedWorker in favor of cookies
Safari and Internet Explorer don't implement the SharedWorker API which
causes the referenced bug.  We're not using the SharedWorker API for
anything other than storing the keystone token.  This change removes the
SharedWorker code altogether, and instead stores the token in a standard
cookie.

We also add a simple setting to the config file which determines whether
the cookie should be marked as "secure".

Change-Id: Ieebcd946b59cb4afb696e12eabf75dec96aeab11
Closes-Bug: 1647590