Refresh node table only once
This change prevents the "flashing" of content in the nodes table as
data is refreshed from the server.

The issue was caused by frequent updates of the "isFetching" property
when making API calls.  We have to make a lot of ajax calls to get the
necessary node information: get the nodes, then get the detail for each
node, and then get the port information for each node.  That's a lot of
API calls.

Now we fetch all node data in a chain, and update the state at the end.

Change-Id: I8049f90b75d02ad2398cbdcc6bf1f03b325bb8e9
Closes-bug: 1649098