[WIP] Add i18n to Notifications
Adding i18n to the Notifications is a little different to other
components because the title and message are defined outside the
component, often coming from external web service calls. This means that
some of the title/message texts passed to the Notification need to be
translated and some might not.

This patch adds some flexibility to the Notification component, so both
message objects as well as plain strings can be used as props for
message and title.

Change-Id: I6335c6ccb92e3077ba6400fa6364446a757863eb