[WIP] Add i18n to Notifications
Adding i18n to the Notifications is a little different to other
components because the title and message are defined outside the
component, often coming from external web service calls. This means that
some of the title/message texts passed to the Notification need to be
translated and some might not.

This patch adds a check to the NotificationToaster component that checks
the type of the message/title properties in the notification state. If
it's a string, it is directly handed over as a prop to the Notfication
component. If it's an i18n message object, it's translated using
intl.formatMessage.

Change-Id: I6335c6ccb92e3077ba6400fa6364446a757863eb