Don't bail webpack by default
bail: true option in webpack.config.js causes webpack to stop building
which is not convenient for development process, as webpack command
then needs to be restarted. For dev we want to keep webpack running so
when problem is fixed, it gets automatically recompiled and browser
refreshed

--bail option should be enough for build
process where we want webpack to fail incase of error.

Change-Id: Id8ce7a2d08b8cd9e3efa5dc51a3eb6cb9ba15efe