Fix up string punctuation
Full sentences should end with a full stop. Sometimes it makes sense
not to e.g. in a header but for things like error messages, I think
it's better to be consistent. Otherwise we end up with needless
duplicates like "There are currently no parameters to configure in this
section." and "There are currently no parameters to configure in this
section".

Also, add a 'description' for a particular string (Add a Plan Name) as
without context it is not obvious what it refers to (e.g. a
button/action vs a form field tooltip).

Change-Id: Iade1dfdf4d251bb85ef8a6f95180cdb90195f065