Update DeploymentPlan page node counts
Nodes counts assigned to Roles are calculated based on actual
'Nodes count' parameters in the plan rather than nodes tagged
to Roles

This change allows GUI to differentiate between tagging Nodes
to Roles and setting counts of Nodes to be deployed as a Role
and avoid syncing problems between those 2 different aspects

Deployment Plan page now displays global number of Nodes
available to distribute among Roles as well as number of assigned
Nodes and number of Nodes available to assign for each specific
Role at given time. Number of available nodes for each Role
depends on tagging applied as well as number of nodes assigned
to other Roles.

This change does not solve Nodes Assignment itself as it is
supposed to be done in a subsequent patch. Nodes assignment still
works under old rules

Partial-Bug: #1640103
Closes-Bug: #1648002
Change-Id: I1f3bfe4276c8c6da48a052497883a7f0b570a157