Fix the deletePlan action
In case of failure of plan deletion, a non-existant function
(planDeleted) was called. This was due to a previous refactor
which renamed the function but didn't fix the invocations.
This patch fixes the issue so that the correct function
(deletePlanFailed) is now called.

Change-Id: I12fcfd6f94b90378055cabc3a65ecb3f012f6f91
(cherry picked from commit 02609106c1e51674adcabe1fa4ac48b6467be0be)