Add puppet to "adding a new language" docs
Non-developers will overwhelmingly install the UI via the "openstack
undercloud" command, which means using puppet-tripleo to manage the UI
configuration file. If we want new languages to be picked up and
available for a normal install, the UI puppet manifest also needs to be
updated as part of enabling a new language.

Change-Id: I46fa539f70552e37109086f377854b9e4da7730c