Fix nodes actions toolbar
An additional span inside the nodes toolbar, added by the i18n
FormattedMessage component, breaks the toolbar by changing the click
event target to the span element; because the span element doesn't
contain the "name" attribute needed to determine the appropriate action.

This patch adds a small method to lookup the closest name attribute
(either on the element or the parent elements), so both buttons with and
without a FormattedMessage of them will work.

Change-Id: I7c565da8bf2f5040ff99b13f02cb0a7f4c07637f
Closes-Bug: #1671100