Remove '\n' linebreak from localised string
This is causing the string to be created without end quote in the pot
file, making it invalid and impossible be upload to Zanata. No new
strings can be added to Zanata until this is resolved.

Note that this doesn't seem to change the tooltip as actually displayed,
as \n doesn't appear to be interpreted as a line break at the moment.

Change-Id: Ic4c07aade661226a6a9cf3106a0dfae0586433a8
Partial-Bug: #1674011