Remove languages with low translations ratios
When a language is translated above 75% on Zanata, it gets
automatically proposed to the project. Once committed, it will only be
removed from git if the translation ratio drops under 40%. However, due
to Story 2000949 this is currently broken for JavaScript translations
(the ratio is calculated on the PO file directly, but we only store the
JSON in git), therefore let's remove under-translated language files
and related configuration manually.

Looking at Zanata [1]:

- Korean is at 35%, however the translator indicated on IRC he will
  work on it this week, so do not remove the files yet.
- German is borderline at 41%, therefore let's remove it as part of
  this patch.
- British English was only used for testing the translation jobs and I
  don't think anyone is actually interested (I've been the only one
  working on it).

If someone wants any of them back, they can do the translations on
Zanata and the bot will submit them back automatically once the ratio
gets back to > 75%.

This is submitted directly to Ocata because it is a released version,
and to give time to translators for working on Pike. If Story 2000949
isn't fixed by the end of Pike, we should do a similar clean-up
manually ahead of the release.

[1] https://translate.openstack.org/iteration/view/tripleo-ui/stable-ocata

Change-Id: If3739e8ee57ce2bd209a842dd7d44caaba139fe6