Include Deployed Nodes in NodesAssignment counts
The nodes available for assigning are all nodes which
are in available state or nodes which are deployed with
currently selected plan. To get this information it is required
to fetch Nova Instances and filter deployed nodes to those
which reference these instances.

This change introduces the selectors and initial instances
state. Subsequent patch should implement actual instances
fetching actions and reducer

Change-Id: I2e33c2b32510a98a189425964e2e6269e5fbf741