Automatically enable all available languages
Right now, we have to create at least two patches for each new language that
becomes available: one to update the tripleo-ui itself, and one to update the
defaults in the puppet code.  This is painful and takes a long time.

This patch turns things around.  Instead of specifying which languages should be
available, we automatically load all of them, and provide users with a way of
disabling some of them should they need to.  When a new language is imported
from the translation team into our code, there is no work necessary.

This is accomplished via Webpack plugin which reads the i18n/locales directory,
parses the data, and generates a Javascript file with the necessary imports and
other useful objects.  The "languages" setting in the config file is changed to
"excludedLanguages", and it's a simple list of language abbreviations.

Closes-Bug: #1682452
Change-Id: Idf5a3314c19be18ca6cabbae1e94bc7cb1d1fe94