Remove linebreak from translatable string
The new line was breaking the translation jobs, as it gets copied as is
into the pot file. In general, it's better to use "+"-style
concatenation for translatable/user-facing strings, as otherwise the
unnecessary spaces from the indentation also get included in the string
to translate, which can be confusing to translators.

Change-Id: I49ba048e1b06199c54b93f2b14ce07b5acdf131d
Closes-Bug: #1691006