Set nodes to manageable before introspecting
Sometimes, nodes can become "available" before they are introspected.  Running
introspection on a node that's in the "available" state will result in an error.
In order to successfully start introspection on a node, the given node must be
in the "manageable" state.  If the user selects a node that's "available" when
choosing which nodes to introspect, we first change its state to "manageable"
before running the introspection request.

This change introduces magic by freeing the user from having to know to set the
node to the correct state.

Change-Id: I77fb9521ae7c03e2fb9668377af4e2e305967886
Closes-Bug: 1639262