Fix plan descriptions on plan cards
After the commit 03b835bcdb1e755cc4ba13ae2b17605ce022328a, plan descriptions
disappeared from plan cards. The code introduced in that commit incorrectly
handles the planEnvs object in PlansActions.js - it treats it as a dictionary
while this objects is an array. This change fixes that so that plan descriptions
are again visible on plan cards.

Change-Id: I053c2a2548085cd497545aa6416a189897c71373
Closes-bug: #1712413