Add version and git sha to debug modal
We add the 'version' and 'gitSha' keys to the AppConfig record, and we
implement the appropriate selectors.  The old webpack version loader is
repurposed into a DefinePlugin which we use to define the VERSION and
GITSHA variables.  Those are read during store hydration, and their
values are stored in the app state.

Change-Id: I9fbef9c62c150aa89bc44bfb8065976a73efde25
Closes-Bug: #1731972