Reenable the undercloud jobs
The undercloud jobs do a sanity check that the UI does work after
installing it via puppet and RPMs, during a regular undercloud
install. The test was set up as part of the basic-ci-check blueprint
[1] (although nowadays the check is probably directly run as part of
quickstart [2]).

This also lets us test that the tripleo-ui RPM gets created correctly.

This job used to run on tripleo-ui but was likely lost during the
migration to zuulv3 once the legacy jobs went away.

[1] https://blueprints.launchpad.net/tripleo/+spec/tripleo-ui-basic-ci-check
[2] https://review.openstack.org/#/c/514795/

Change-Id: I8fcb7c7d96c13101c00c5427092ae2d7503f65ca