Stop logging over websocket on logout
When a user logs out of the application, we correctly close the
websocket connection.  However, the logging mechanism still tries to
send messages over the closed connection.  This commit sets
`authenticated` value in the `logger` state to `false` upon logout in
addition to closing the websocket connection.  This informs the logging
system that no more messages should be sent.

Change-Id: Iaf365d78cf5e2b35c4f035a9f068266131940553
Closes-Bug: 1746583