Disable plan upload/edit button during transition
The "Upload Files and Create/Edit Plan" button is currently not being
disabled after it has been clicked and while the create/update process
is still running. Clicking it twice results in an error and should not
be possible in the first place.

This patch disables the button during a transition by checking the
`isTransitioningPlan` prop in addition to the `canSubmit` prop.

Change-Id: I4324ca828ff2ea1eccafa8d7faae7a181b736b2c
Closes-Bug: #1747932