Discoverable Ironic Drivers
This patch wires in the ability to discover which drivers are active
in Ironic (via query to /drivers) and only displays the ability to
configure drivers which are currently enabled in Ironic.

Three extra driverFields were also added in this patch. (These are
currently static so that we can easily support language customizations
for the UI). The new additions include Ilo, Redfish, and Iboot.
With the additions of Ilo and Redfish the UI would now support the
same stock settings as our Undercloud with support for Ilo, Redfish,
IPMI, and DRAC.

Co-Authored-By: Jiri Tomasek <jtomasek@redhat.com>
Co-Authored-By: Honza Pokorny <honza@redhat.com>

Change-Id: Ic5693506e67ca1a54ba88bdb4e6a02be0ebe4644