Update Manage Roles link to be Inline Button
This patch changes the Manage Roles link on the deployment plan
page to be a button. It also moves it to be inline with all
other main buttons on the page.

Note that the Save and Cancel buttons when changing the number
on nodes assigned to roles will remain floating to the right.

Change-Id: I8e5069c276e49d810554db81d71c06e62223fc56
Closes-Bug: #1757014