Better tarball upload validation
When creating or updating plans via a tarball, the validation
we had so far only looked at the file extension to determine the
validity of the provided file. As the file extension can be spoofed
by simply renaming the file, a better way is to determine the real
MIME type by looking at the file signature[1]. This patch implements
that change.

[1] https://en.wikipedia.org/wiki/File_signature

Change-Id: Idd3f4e7f83f2ff03fa8c90373acac4d6828ed81e