Remove h1 page header on deployment plan page
This patch removes the h1 plan name at the top of the deployment
plan details page. Since the plan name has been included in
the breadcrumb, this large page title isn't needed. Removing it
will save precious vertical space and align with PatternFly
standards.

Making this change required updating the css overrides to include
a border-bottom on the breadcrumb class since this previously was
on the page header. Also, margin is set to 0 to fix spacing.

Change-Id: I24429e8c343b97f04ef5de0f5efdbf5b1f16663b
Closes-Bug: #1757540