Upgrade webpack to 4.5.0
Earlier versions of webpack had a bug in the BannerPlugin which we use
to insert the TripleO license header into production builds of the app.
The license header was treated as a normal comment, and removed by the
minification mechanism.  Upgrading to a more recent version of webpack
fixes the problem.

Change-Id: I91bf6a455600a80dd2cdd8abb46451642d523bc6
Closes-Bug: 1762375