Mask Passwords and allow Copy to Clipboard
This patch masks the IPMI password in the register nodes form
along with the Deployment Admin Password on a successful
deployment.

This also adds a 'Copy' button next to the Admin Password that
allows copying the password to the clipboard for future use.

Additionally, this patch allows the user to show/hide the Admin
Password.

NOTE: This adds two dependencies to our package.json. Both are
under the MIT License:
1.https://www.npmjs.com/package/react-copy-to-clipboard
2.https://www.npmjs.com/package/react-password-mask

I need some help on the following if anyone wants to jump in:
1. I don't know how to make use of the PasswordMask component
for just the input on the IPMI password field. Currently it's
being rendered using the Field component and accepting
HorizontalInput as it's component type which renders both label
and input. Any suggestions on a way to insert just PasswordMask
as the input component are welcome.

Change-Id: I1f04278ece8c4e3de8b09ef8eaa51c6ad4f9421f