Mask Passwords and allow Copy to Clipboard
This patch masks the IPMI password in the register nodes form
along with the Deployment Admin Password on a successful
deployment.

This also adds a 'Copy' button next to the Admin Password that
allows copying the password to the clipboard for future use.

Additionally, this patch allows the user to show/hide the Admin
Password.

NOTE: This adds two dependencies to our package.json. Both are
under the MIT License:
1.https://www.npmjs.com/package/react-copy-to-clipboard
2.https://www.npmjs.com/package/react-password-mask

I need some help on the following if anyone wants to jump in:
1. It seems that the react-password-mask package is giving me
some dependency errors with react, but I think our version
is within the guidelines. Needs more troubleshooting and a
possible suggestion for fixing this issue.
2. I don't know how to make use of the PasswordMask component
for just the input on the IPMI password field. Currently it's
being rendered using the Field component and accepting
HorizontalInput as it's component type which renders both label
and input. Any suggestions on a way to insert just PasswordMask
as the input component are welcome.
3. I'm not sure the best way to get the password that is now
rendered via the PasswordMask component to stay inline with
the 'Password:' label.

Change-Id: I1f04278ece8c4e3de8b09ef8eaa51c6ad4f9421f