Mask Passwords and allow Copy to Clipboard
This patch masks the IPMI password in the register nodes form
along with the Deployment Admin Password on a successful
deployment.

This also adds a 'Copy' button next to the Admin Password that
allows copying the password to the clipboard for future use.

NOTE: The copy to clipboard depends on installing
'react-copy-to-clipboard'.

Change-Id: I1f04278ece8c4e3de8b09ef8eaa51c6ad4f9421f