Update plans state management
Reorganize plans fetching into separate actions which are
separately invoked when needed PlansList lists the plans, PlanCard
loads plan details - plan-environment.yaml to fetch the description.
Plans store is reorganized apropriately, this leads to simpler
reducer as there are no collisions between plans and plan details
loading.

Prepare EditPlan and EditPlanForm as plan editting is currently
broken, the form is going to be fixed in subsequent patch

Plans transitions are handled as a List of transitions

When plan is created, it is automatically added to plans list
and newly mounted PlanCard takes care of fetching the details.
No plans refetch is required

Story: 2003188
Task: 23338
Change-Id: Ie4c1fc256bbe3243e07edb03d7c60804016d52bd