Use swiftUrl as a base for exported plan download URL
This changes the way the exported plan and logs download URL is generated.
The temporary url is parsed, path is stripped from API version
and AUTH parts and appended with search section to swiftUrl.

This ensures that generated URL is always accessible since it is based
on swiftUrl.

Story: 2001869
Task: 14323

Change-Id: I3961eba69cb910e0fb63ca763f301bfc4ba7c05c