Move tripleo-ui out of the shared tripleo gate
So we don't actually rely on the tripleo-ui for the tripleo jobs other
than to be installable in a container on the udnercloud.  The
likelihood of anything being merged in tripleo-ui that would hang the
gate is very small and likely to be caught by the existing CI.  This
change moves the tripleo-ui out of the tripleo gate queue to it's own
tripleo-ui gate.  This will allow tripleo-ui to merge things outside of
the long tripleo queue

Change-Id: I25dfabad14ae816692e1093bc184b28fa239b20e